+++ src/main/java/techcourse/myblog/service/ArticleService.java
@@ -5,9 +5,9 @@ import java.util.List;
 import techcourse.myblog.domain.article.Article;
 import techcourse.myblog.domain.article.Contents;
 import techcourse.myblog.domain.user.User;
-import techcourse.myblog.exception.NotFoundArticleException;
-import techcourse.myblog.exception.UnauthorizedException;
 import techcourse.myblog.repository.ArticleRepository;
+import techcourse.myblog.service.exception.NotFoundArticleException;
+import techcourse.myblog.service.exception.UnauthorizedException;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
