+++ src/main/java/techcourse/myblog/service/exception/NotFoundArticleException.java
@@ -1,4 +1,4 @@
-package techcourse.myblog.exception;
+package techcourse.myblog.service.exception;
 
 public class NotFoundArticleException extends RuntimeException {
 	public NotFoundArticleException() {
