+++ src/main/java/techcourse/myblog/service/exception/NotFoundUserException.java
@@ -1,4 +1,4 @@
-package techcourse.myblog.exception;
+package techcourse.myblog.service.exception;
 
 public class NotFoundUserException extends RuntimeException {
 	public NotFoundUserException() {
