+++ src/main/java/techcourse/myblog/service/exception/NotMatchPasswordException.java
@@ -1,4 +1,4 @@
-package techcourse.myblog.exception;
+package techcourse.myblog.service.exception;
 
 public class NotMatchPasswordException extends RuntimeException {
 	public NotMatchPasswordException() {
