+++ src/main/java/techcourse/myblog/service/exception/UnauthorizedException.java
@@ -1,4 +1,4 @@
-package techcourse.myblog.exception;
+package techcourse.myblog.service.exception;
 
 public class UnauthorizedException extends RuntimeException {
 	public UnauthorizedException() {
