+++ src/main/java/techcourse/myblog/service/request/UserSignUpInfoDto.java
@@ -1,4 +1,4 @@
-package techcourse.myblog.dto.request;
+package techcourse.myblog.service.request;
 
 import javax.validation.constraints.Email;
 import javax.validation.constraints.NotBlank;
