+++ src/main/resources/application.properties
@@ -1,8 +1,8 @@
 spring.devtools.restart.enabled=true
 spring.devtools.livereload.enabled=true
 
-spring.jpa.hibernate.ddl-auto=create-drop
-spring.datasource.url=jdbc:mysql://localhost:3306/jwp_blog?serverTimezone=UTC
+spring.jpa.hibernate.ddl-auto=create
+spring.datasource.url=jdbc:mysql://localhost:3306/jwp_blog?serverTimezone=UTC&serverTimezone=Asia/Seoul
 spring.datasource.username=blog
 spring.datasource.password=blog123
